Birds: Catalog.zip: Update 2024/06/07: Includes atomic date fields.
https://umlsait.atlassian.net/browse/LMD-301

Requested by Brett Benz:

For studies of birds on their breeding range, the researcher needs a date range search defined by start and end month, but independent of the year. Bird collections from breeding, migratory, and overwintering periods are distinct, and a search functionality is needed that can isolate one of these groups while excluding the others. For example, searching for records with a collection event from June to August, across all years. This functionality is supported within Specify 6 and 7 but not in the Specify Web Portal.